backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
supabase>=2.3.0

# Audio Processing
numpy>=1.26.0
librosa>=0.10.1
soundfile>=0.12.1
praat-parselmouth>=0.4.3

# Speech-to-Text
openai-whisper

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Additional dependencies for whisper
torch
torchaudio
